+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+v0.0.2 - 2019-09-24
+
+* Added `parquet-to-text-stream` supporting `csv` format
+
 v0.0.1 - 2019-09-23
 -------------------
 
-* Initial release
+* Initial release, with `parquet-to-arrow-slice`